requirements.txt
gunicorn==21.2.0
flask==2.3.3
flask-cors==4.0.0
werkzeug==2.3.7
openai==0.28.1
python-dotenv==1.0.0
redis==5.0.1
celery==5.3.4

